Handwritten_character_recognition/requirements.txt
# Core dependencies
tensorflow~=2.19.0
tensorflow-datasets
numpy~=2.1.3
Pillow~=11.3.0
opencv-python~=4.12.0.88
flask~=3.1.1
gunicorn
flask_cors~=6.0.1
kagglehub~=0.3.12
scipy~=1.16.0
matplotlib~=3.10.3
pandas~=2.3.1
scikit-learn~=1.7.1